refactor : Refactor StringAddCalculator
* Create StringSplitter
