+++ src/main/java/StringAddCalculator.java
@@ -1,15 +1,9 @@
-import java.util.Arrays;
 import java.util.List;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 import static java.util.stream.Collectors.collectingAndThen;
 import static java.util.stream.Collectors.toList;
 
 public class StringAddCalculator {
-    private static final String REGEX_FOR_DEFAULT = ",|:";
-    private static final String REGEX_FOR_CUSTOM = "//(.)\n(.*)";
-
     public static int addString(String input) {
         if (isNullOrEmpty(input)) {
             return 0;
@@ -26,22 +20,10 @@ public class StringAddCalculator {
     }
 
     private static PositiveNumbers convertFromStringToNumbers(String input) {
-        List<String> strings = splitByCustom(input);
+        List<String> strings = StringSplitter.split(input);
         return covertToNumber(strings);
     }
 
-    private static List<String> splitByDefault(String input) {
-        return Arrays.asList(input.split(REGEX_FOR_DEFAULT));
-    }
-
-    private static List<String> splitByCustom(String input) {
-        Matcher m = Pattern.compile(REGEX_FOR_CUSTOM).matcher(input);
-        if (m.find()) {
-            String customDelimiter = m.group(1);
-            return Arrays.asList(m.group(2).split(customDelimiter));
-        }
-        return splitByDefault(input);
-    }
 
     private static PositiveNumbers covertToNumber(List<String> stringNumbers) {
         try {
